add reconcile and delete op
