+++ .ci/component_descriptor
@@ -15,6 +15,8 @@ echo "> building component gardenlogin-controller-manager"
 CA_PATH="$(mktemp -d)"
 mv "$COMPONENT_DESCRIPTOR_PATH" "$CA_PATH/component-descriptor.yaml"
 
-component-cli ca "${CA_PATH}" "${CTF_PATH}" \
+echo "CA_PATH ${CA_PATH}"
+
+component-cli component-archive "${CA_PATH}" "${CTF_PATH}" \
     -r  "$SOURCE_PATH"/.landscaper/resources.yaml \
     VERSION="${EFFECTIVE_VERSION}"
