+++ .landscaper/blueprint/config
@@ -1,0 +1,1 @@
+../../config
