+++ .landscaper/blueprint/gardenlogin-deploy-execution.yaml
@@ -1,9 +1,6 @@
 deployItems:
 - name: gardenlogin-container-deployer
   type: landscaper.gardener.cloud/container
-#  target:
-#    name: {{ index .imports "cluster" "metadata" "name" }}
-#    namespace: {{ index .imports "cluster" "metadata" "namespace" }}
   config:
     apiVersion: container.deployer.landscaper.gardener.cloud/v1alpha1
     kind: ProviderConfiguration
