+++ .landscaper/container/go.mod
@@ -19,6 +19,7 @@ require (
 	k8s.io/client-go v11.0.1-0.20190409021438-1a26190bd76a+incompatible
 	k8s.io/component-base v0.20.7
 	sigs.k8s.io/controller-runtime v0.8.3
+	sigs.k8s.io/yaml v1.2.0
 )
 
 replace k8s.io/client-go => k8s.io/client-go v0.20.2
