+++ .landscaper/container/hack/create-installation.sh
@@ -11,12 +11,8 @@ TMP_DIR="$(mktemp -d)"
 INSTALLATION_PATH="${TMP_DIR}/installation.yaml"
 CONFIGMAP_PATH="${TMP_DIR}/configmap.yaml"
 
-IMAGE_REGISTRY="${IMAGE_REGISTRY:-eu.gcr.io/gardener-project/development}"
-
-endpointData=$(echo "${APPLICATION_CLUSTER_ENDPOINT}" | base64 -w0)
-multiClusterData=$(echo "true" | base64 -w0)
-namePrefixData=$(echo "terminal-" | base64 -w0)
-namespaceData=$(echo "terminal-system" | base64 -w0)
+IMAGE_REGISTRY="${IMAGE_REGISTRY:-eu.gcr.io/gardener-project/development/images}"
+CD_REGISTRY="${CD_REGISTRY:-eu.gcr.io/gardener-project/development}"
 
 cat << EOF > ${INSTALLATION_PATH}
 apiVersion: landscaper.gardener.cloud/v1alpha1
@@ -28,7 +24,7 @@ spec:
     ref:
       repositoryContext:
         type: ociRegistry
-        baseUrl: ${IMAGE_REGISTRY}
+        baseUrl: ${CD_REGISTRY}
       componentName: github.com/gardener/gardenlogin-controller-manager/.landscaper/container
       version: ${EFFECTIVE_VERSION}
 
@@ -38,9 +34,9 @@ spec:
 
   imports:
     targets:
-    - name: application-cluster-target
+    - name: applicationClusterTarget
       target: "#application-cluster"
-    - name: runtime-cluster-target
+    - name: runtimeClusterTarget
       target: "#runtime-cluster"
     data:
     - name: applicationClusterEndpoint
@@ -64,6 +60,11 @@ spec:
 
 EOF
 
+endpointData="${APPLICATION_CLUSTER_ENDPOINT}"
+multiClusterData=$(echo "true" | base64 -w0)
+namePrefixData="gardenlogin-"
+namespaceData="gardenlogin-system"
+
 cat << EOF > ${CONFIGMAP_PATH}
 apiVersion: v1
 kind: ConfigMap
@@ -71,9 +72,10 @@ metadata:
   name: gardenlogin-container-deployer
 data:
   applicationClusterEndpoint: ${endpointData}
-  multiClusterDeploymentScenario: ${multiClusterData}
   namePrefix: ${namePrefixData}
   namespace: ${namespaceData}
+binaryData:
+  multiClusterDeploymentScenario: ${multiClusterData}
 
 EOF
