--- .landscaper/container/pkg/api/exports.go
@@ -1,10 +1,0 @@
-// Copyright (c) 2021 SAP SE or an SAP affiliate company. All rights reserved. This file is licensed under the Apache Software License, v. 2 except as noted otherwise in the LICENSE file
-// SPDX-FileCopyrightText: 2021 SAP SE or an SAP affiliate company and Gardener contributors
-//
-// SPDX-License-Identifier: Apache-2.0
-
-package api
-
-// Exports defines the structure for the exported data which might be consumed by other components.
-type Exports struct {
-}
