+++ config/manager/manager.yaml
@@ -63,7 +63,6 @@ spec:
         volumeMounts:
           - mountPath: /etc/gardenlogin-controller-manager
             name: manager-config
-      serviceAccountName: controller-manager
       terminationGracePeriodSeconds: 10
       volumes:
         - configMap:
